+++ lib/mirror.dart
@@ -1,4 +1,4 @@
-/// Mirror classes used by the ContainerBuilder.
+/// Mirror classes used by the ScopeBuilder.
 library mirror;
 
 export 'package:needle/src/mirrors.dart';